feat: Created DTOs for password recovery, reset, and login; refactored classes and fixed application logic and annotations; performed testing using Swagger and Postman; implemented pagination in updated classes, services, and controllers to comply with new rules; also added logging to the package.

+++ library-backend/src/main/java/com/altis/library/books/services/BooksServices.java
@@ -1,30 +1,39 @@
 package com.altis.library.books.services;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import com.altis.library.books.models.dtos.BooksRequestDTO;
+import com.altis.library.books.models.dtos.BooksResponseDTO;
 import com.altis.library.books.models.dtos.UpdateRequestDTO;
 import com.altis.library.books.models.entities.books;
 import com.altis.library.books.repositories.BooksRepository;
 import com.altis.library.publishers.models.entities.publishers;
 import com.altis.library.publishers.repositories.PublishersRepository;
+import com.altis.library.rents.repositories.RentsRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
+import java.time.LocalDate;
 
 @Service
 public class BooksServices {
 
     private final BooksRepository booksRepository;
     private final PublishersRepository publishersRepository;
+    private final RentsRepository rentsRepository;
 
     public BooksServices(
             BooksRepository booksRepository,
-            PublishersRepository publishersRepository
+            PublishersRepository publishersRepository,
+            RentsRepository rentsRepository
     ) {
         this.booksRepository = booksRepository;
         this.publishersRepository = publishersRepository;
+        this.rentsRepository = rentsRepository;
     }
 
-    public books create(BooksRequestDTO bookData) {
+    @Transactional
+    public BooksResponseDTO create(BooksRequestDTO bookData) {
 
         if (booksRepository.existsByIsbn(bookData.isbn())) {
             throw new RuntimeException("ISBN already registered");
@@ -40,20 +49,33 @@ public class BooksServices {
 
         book.setTitle(bookData.title());
         book.setIsbn(bookData.isbn());
-        book.setPublicationDate(bookData.publicationDate());
+        book.setPublicationDate(LocalDate.now());
         book.setGenre(bookData.genre());
         book.setNumberPages(bookData.numberPages());
         book.setQuantity(bookData.quantity());
-
         book.setPublisher(publisher);
 
-        return booksRepository.save(book);
+        books savedBook = booksRepository.save(book);
+
+        return new BooksResponseDTO(
+                savedBook.getId(),
+                savedBook.getTitle(),
+                savedBook.getPublicationDate(),
+                savedBook.getGenre(),
+                savedBook.getNumberPages(),
+                savedBook.getQuantity(),
+                savedBook.getPublisher().getId(),
+                savedBook.getCreateDt(),
+                savedBook.getUpdateDt()
+        );
     }
 
-    public List<books> findAll() {
-        return booksRepository.findAll();
+    @Transactional
+    public Page<books> findAll(Pageable pageable) {
+        return booksRepository.findAll(pageable);
     }
 
+    @Transactional(readOnly = true)
     public books findById(Long id) {
 
         return booksRepository.findById(id)
@@ -62,6 +84,7 @@ public class BooksServices {
                 );
     }
 
+    @Transactional
     public books update(Long id, BooksRequestDTO bookData) {
 
         books book = findById(id);
@@ -73,8 +96,6 @@ public class BooksServices {
                 );
 
         book.setTitle(bookData.title());
-        book.setIsbn(bookData.isbn());
-        book.setPublicationDate(bookData.publicationDate());
         book.setGenre(bookData.genre());
         book.setNumberPages(bookData.numberPages());
         book.setQuantity(bookData.quantity());
@@ -83,6 +104,7 @@ public class BooksServices {
         return booksRepository.save(book);
     }
 
+    @Transactional
     public books partialUpdate(
             Long id,
             UpdateRequestDTO bookData
@@ -92,10 +114,6 @@ public class BooksServices {
 
         if (bookData.title() != null) {
             book.setTitle(bookData.title());
-        }
-
-        if (bookData.publicationDate() != null) {
-            book.setPublicationDate(bookData.publicationDate());
         }
 
         if (bookData.genre() != null) {
@@ -124,9 +142,16 @@ public class BooksServices {
         return booksRepository.save(book);
     }
 
+    @Transactional
     public void delete(Long id) {
 
         books book = findById(id);
+
+        if (rentsRepository.existsByBookId(id)) {
+            throw new RuntimeException(
+                    "Book cannot be deleted because it has registered rents"
+            );
+        }
 
         booksRepository.delete(book);
     }
